# swagger/swagger-v1.0.2.yaml
openapi: 3.0.3
info:
  title: Liga Nunez
  description: |- 
    HTTP requests y responses documentadas
    
    Links de utilidad:
    - [Requerimientos](https://docs.google.com/document/d/1_kFwMBbmoDiY1nCElglinTN5vLZckk4n5RgaN-1zyKg/edit?pli=1)
    - [Tablero JIRA](https://venedicto.atlassian.net/jira/software/projects/LGN/boards/9)
  version: 1.0.2
servers:
  - url: http://localhost:8080/api/ln/
...
paths:
...
  /tournament:
    post:
      tags:
       - tournament
      summary: Creación de torneo
      operationId: createTournament
      parameters:
        - name: token
          in: header
          description: Token de autenticación del usuario
          required: false
          schema:
            type: string
      requestBody:
        description: Información del torneo (sin id ni código)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament'
      responses:
        '201':
          description: Torneo creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '400':
          description: Error en la request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
    delete:
      tags:
       - tournament
      summary: Eliminación de torneo
      operationId: deleteTournament
      parameters:
        - name: token
          in: header
          description: Token de autenticación del usuario
          required: false
          schema:
            type: string
        - name: tournamentId
          in: header
          description: ID del torneo
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Torneo eliminado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Torneo inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
  /tournament/{id}:
    get:
      tags:
        - tournament
      summary: Carga de torneo
      operationId: getTournamentData
      parameters:
        - name: id
          in: path
          description: ID del torneo
          required: true
          schema:
            type: string
        - name: code
          in: header
          description: Código de acceso ingresado
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Información obtenida con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTournamentHttpResponse'
        '400':
          description: Error en la request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Torneo inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
    put:
      tags:
       - tournament
      summary: Actualización del torneo
      operationId: updateTournament
      parameters:
        - name: token
          in: header
          description: Token de autenticación del usuario
          required: false
          schema:
            type: string
        - name: id
          in: path
          description: ID del torneo
          required: true
          schema:
            type: string
      requestBody:
        description: Información del torneo (sin id ni código)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tournament'
      responses:
        '200':
          description: Torneo actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '400':
          description: Error en la request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '404':
          description: Torneo inexistente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '500':
          description: Error del servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
        '503':
          description: Base de datos no disponible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpResponse'
...
components:
  schemas:
    Tournament:
      type: object
      properties:
        id:
          type: string
          example: "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
        code:
          type: integer
          example: 456509
        name:
          type: string
          example: "Torneo 1"
        type:
          type: string
          example: M
          enum: [F,M]
...
    Error:
      type: object
      properties:
        code:
          type: string
          example: "P01"
        description:
          type: string
          example: "Error interno"
    HttpResponse:
      type: object
      properties:
        opCode:
          type: string
          example: "404"
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
...
    GetTournamentHttpResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/HttpResponse'
      properties:
        data:
          $ref: '#/components/schemas/Tournament'
...